# openapi.yaml
openapi: 3.1.0
info:
  title: Universal file converter
  version: 2.0.0
  summary: Converts documents, spreadsheets, presentations and images using LibreOffice headless.
  description: |
    Uploads a file and receives it converted to another format.

    ## The contract

    This API is consumed by two clients: a **shipped Android app** and a
    browser-based **converter web app**. The Android app is why the following
    properties are part of the contract and are not free to change - a mobile
    client cannot be patched the moment the server changes, the way the web
    app can be redeployed alongside it. The web app relies on the JSON error
    envelope and the status-before-body rule the same way, but carries no
    separate abort timer or media-type check of its own worth pinning here.

    > **Breaking change: `POST /convert` without a target no longer exists.**
    > Clients must post to `/convert/pdf`. A client still using the bare path
    > gets the catch-all `404`, whose message tells the user to update the app.

    - **Success is `200` with the target's own media type.** For `pdf` that is
      exactly `application/pdf`, with no charset suffix: the client compares the
      header and refuses anything else, so a stray `text/html` on a `200` is a
      client-visible failure.
    - **Failure is a non-2xx with the JSON error envelope.** The client reads
      `error.message` and shows it verbatim in a dialog, falling back to
...
      downstream.
...

...

...

...
    error.

...

...

...

...

...
servers:
...
paths:
  /convert/{target}:
    post:
      tags: [convert]
      operationId: convertToFormat
      summary: Convert one or more files to a named format
      description: |
        Converts every uploaded file to `target`, which must be an id from
        `GET /formats` (`pdf`, `odt`, `docx`, `txt`, `html`, `rtf`, `epub`,
        `ods`, `xlsx`, `csv`, `odp`, `pptx`, `png`, `jpg`, `tables`, `layers`,
        `pdfa`), independently of one another - a file that cannot reach
        `target` (wrong extension, damaged document) does not stop the rest
        of the batch when more than one file was uploaded. With exactly one
        file, that failure IS the response, exactly as it always was.

        Each document is parsed by LibreOffice, which is a large C++ codebase
        handling untrusted input. **Treat a conversion failure as an expected,
        normal event**, not as an incident.

        The endpoint is unauthenticated and cannot be otherwise: the client is
        an APK, so any secret shipped inside it is public. Requests are
        rate-limited per IP and conversions run under a bounded concurrency
        limit; both are sources of `429` and `503` that a well-formed request
        can hit.

        A PDF with no extractable text at all (a scan), or one built on
        Type3 fonts, asking for `docx` is OCR'd before reconstruction by
        default - see the `ocr` field on the request body.
      x-max-upload-bytes: 104857600
      parameters:
        - name: target
          in: path
          required: true
          description: |
            The format to produce. Case-sensitive, without a leading dot.
          schema:
            $ref: '#/components/schemas/TargetId'
      requestBody:
        $ref: '#/components/requestBodies/Upload'
      responses:
        '200':
          $ref: '#/components/responses/ConvertedToTarget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/UnknownTarget'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedTarget'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ConvertFailed'
        '503':
          $ref: '#/components/responses/Busy'
        '504':
          $ref: '#/components/responses/Timeout'
...
components:
  requestBodies:
    Upload:
      required: true
      description: |
        One or more file parts, all named `files`, plus the optional `ocr`
        field below. A request with no `files` part is rejected with `400`;
        one under any other name is simply not this field and is ignored the
        same way any unrecognised part would be.

        Each file is capped at **104857600 bytes (100 MB)**, matching the
        client's own `MAX_UPLOAD_BYTES` and the reverse proxy's body limit -
        exceeding it at any layer produces `413`. With more than one file,
        their combined size is additionally capped (`MAX_CONVERT_TOTAL_BYTES`,
        100 MB by default) and their count is capped at `MAX_CONVERT_FILES`
        (15 by default), both also `413`.
      content:
        multipart/form-data:
          schema:
            type: object
            required: [files]
            properties:
              files:
                type: array
                minItems: 1
                items:
                  type: string
                  format: binary
                description: |
                  The file(s) to convert. Each one's **filename extension**
                  selects its own import filter and must be one of the
                  extensions listed by `GET /formats` (case-insensitive). The
                  declared `Content-Type` of each part is ignored.

                  Filenames are never used as a path on disk: each upload is
                  written to a server-generated name inside its own temporary
                  directory, so a name like `../../etc/passwd.docx` is
                  harmless.
              ocr:
                type: string
                enum: ['true', 'false']
                default: 'true'
                description: |
                  Only meaningful for a PDF asking for `docx` with no
                  extractable text (a scan) or built on Type3 fonts -
                  silently ignored by every other source/target pair. When
                  true (the default), the page is OCR'd, so it becomes real,
                  reflowable text instead of an uneditable picture of one -
                  with no embedded image, for either case. Set to `false` to
                  skip OCR and get the page as a plain image with no text
                  instead. Applies to every file in the request alike.
          encoding:
            files:
              contentType: application/octet-stream
...
  schemas:
...
    TargetId:
      type: string
      description: |
        Identifier of an output format, as used in the `/convert/{target}` path
        and returned by `GET /formats`.
      enum:
        [pdf, odt, docx, txt, html, rtf, epub, ods, xlsx, csv, odp, pptx, png, jpg, tables, layers, pdfa, markdown, zip, tar, "tar.gz", "tar.bz2", "7z", cbz, srt, vtt, ass, ssa, tsv, json, yaml, jsonl, bmp, gif, tiff, webp, avif, ico, "png-image", "jpg-image", svg, heic, heif, emf, wmf, eps, jxl, jp2, qoi, tga, pcx, apng, xml, toml, ini, sqlite, "tar.zst", obj, stl, ply, glb, "3mf", mobi, azw3, fb2, lrf, pdb, snb, kepub, ttf, otf, woff, woff2, parquet, orc, feather]
...
  responses:
...
    ConvertedToTarget:
      description: |
        The converted file - or files.

        **Exactly one file uploaded:** the media type is the target's own -
        `application/pdf` for `pdf`, `text/csv` for `csv`, and so on. The image
        targets are the exception: `png` and `jpg` always answer with
        `application/zip`, because there is one image per page and a ZIP is
        the only way to put several files in one response.

        **Two or more files uploaded:** always `application/zip`, regardless
        of target, holding one entry per input file named
        `<NN>-<original name>.<target extension>` (`01-report.pdf`,
        `02-invoice.pdf`, ...), or a `<NN>-<original name>/` folder of images
        for an image target. A file that failed to convert contributes no
        entry of its own; instead, if any file in the batch failed, the
        archive additionally holds `errors.json` - an array of
        `{ file, code, message }`, in upload order, naming every failure. A
        batch where every file failed still answers `200` with a ZIP holding
        only `errors.json`, because the failures are per file, not a fact
        about the request as a whole.
      headers:
        X-Request-Id:
          schema:
            type: string
        Content-Disposition:
          description: |
            The uploaded file's name with the target's extension. Sanitised:
            path segments and control characters are stripped, and a name with
            nothing usable in it becomes `converted.<ext>`.

            Sent in both RFC 6266 forms - a quoted ASCII `filename` for older
            clients and a percent-encoded UTF-8 `filename*` for current ones -
            so a non-ASCII name survives.
          schema:
            type: string
          example: attachment; filename="Quarterly report.xlsx"; filename*=UTF-8''Quarterly%20report.xlsx
      content:
        application/pdf:
          schema:
            type: string
            format: binary
        application/zip:
          schema:
            type: string
            format: binary
            description: |
              A ZIP, for every target that reports `multiple: true` at
              `GET /formats`. What is inside depends on the target:

                - `png` / `jpg` - one image per page, named `slide-1.png`,
                  `slide-2.png`, ...
                - `layers` - one PNG per layer of a PSD, at the path its group
                  gives it (`Buttons/Hover.png`), plus `manifest.json`
                  describing every layer, including any that produced no image
                  and why.

              The response is an archive even when it holds a single file, so
              that the content type never depends on how many pages or layers
              the upload happened to have.
        application/vnd.oasis.opendocument.text:
          schema:
            type: string
            format: binary
        application/vnd.openxmlformats-officedocument.wordprocessingml.document:
          schema:
            type: string
            format: binary
        application/vnd.oasis.opendocument.spreadsheet:
          schema:
            type: string
            format: binary
        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
          schema:
            type: string
            format: binary
        application/vnd.oasis.opendocument.presentation:
          schema:
            type: string
            format: binary
        application/vnd.openxmlformats-officedocument.presentationml.presentation:
          schema:
            type: string
            format: binary
        application/epub+zip:
          schema:
            type: string
            format: binary
        application/rtf:
          schema:
            type: string
            format: binary
        text/plain:
          schema:
            type: string
        text/csv:
          schema:
            type: string
        text/html:
          schema:
            type: string
    BadRequest:
      description: |
        The request was not a usable upload: no file part named `file`, several
        files, a part under a different name - or, distinctly, the optional
        `ocr` field was something other than `true`/`false`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            badRequest:
              summary: No usable file part.
              value:
                error:
                  code: E_BAD_REQUEST
                  message: The document could not be received. Please try again.
            invalidOcrField:
              summary: The `ocr` field was not `true`/`false`.
              value:
                error:
                  code: E_INVALID_FIELD
                  message: The "ocr" field must be "true" or "false".
...
    UnknownTarget:
      description: |
        The target in the path is not a format this service can produce at all -
        `/convert/banana`. Distinct from a real target the source cannot become,
        which is a `415`: this one is a `404` because the address does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_UNKNOWN_TARGET
              message: 'That is not a format this converter can produce. Available: PDF, ODT, DOCX, TXT, HTML, RTF, EPUB, ODS, XLSX, CSV, ODP, PPTX, PNG, JPG, XLSX (tables), PNG (layers), PDF/A, Markdown, ZIP, TAR, TAR.GZ, TAR.BZ2, 7Z, CBZ, SRT, VTT, ASS, SSA, TSV, JSON, YAML, JSONL, BMP, GIF, TIFF, WEBP, AVIF, ICO, PNG (image), JPG (image), SVG, HEIC, HEIF, EMF, WMF, EPS, JXL, JPEG 2000, QOI, TGA, PCX, APNG, XML, TOML, INI, SQLite, TAR.ZST, OBJ, STL, PLY, GLB, 3MF, MOBI, AZW3, FB2, LRF, PDB, SNB, KEPUB, TTF, OTF, WOFF, WOFF2, Parquet, ORC, Feather.'
    TooLarge:
      description: |
        The upload exceeds 104857600 bytes (100 MB). Applied by the reverse proxy,
        by the multipart parser and by the client, all using the same limit.

        Also returned when an image target would produce more pages than the
        server will hold in memory at once, which is a bound on the archive
        rather than on the upload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_TOO_LARGE
              message: This document is too large to convert.
...
    UnsupportedTarget:
      description: |
        A real target, but not one this source can become - asking for `png`
        from a `.docx`, say. This is a `415` and not a `404`, because the
        address exists; it is the combination that does not.

        The message names what the document CAN become instead, because the
        useful answer to "can I have this as X" is "no, but here is what you can
        have". The list varies by source, so the example below is the one for a
        Word document; see `GET /formats` for the rest.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_UNSUPPORTED_TARGET
              message: 'A .docx file can be converted to: PDF, ODT, TXT, HTML, RTF, EPUB, XLSX (tables).'
    Unprocessable:
      description: |
        The document was received and read, and the answer is that there is no
        answer. Two quite different situations land here, and both are the
        document's own doing rather than a fault:

        **Password protected.** Detected before LibreOffice runs, because
        LibreOffice reports an encrypted document the same way it reports a
        corrupt one, and telling a user their file is damaged when it merely
        needs a password is both wrong and unhelpful.

        **No tables, and no layers.** Only an `extract` target can produce
        either: those are the targets that read the document's contents rather
        than converting the document as a whole, so they are the only ones that
        can open a file successfully and find nothing to work with. Reporting
        one as a conversion failure would say the file may be damaged when the
        truth is that it is a perfectly good document with no tables in it, or
        one whose layers are all adjustment and text layers and therefore have
        no pixels to write.

        **Not a table.** A `csv`/`tsv`/`jsonl` request against a JSON/YAML
        source whose content is not shaped for that target - a single object
        rather than a top-level array, or nesting a delimited format has no
        way to represent. The same reasoning as no-tables/no-layers: the
        document is valid, just not shaped for what was asked of it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            encrypted:
              summary: The document needs a password.
              value:
                error:
                  code: E_ENCRYPTED
                  message: This document is password protected.
            noTables:
              summary: A `tables` request on a document with no tables.
              value:
                error:
                  code: E_NO_TABLES
                  message: This document does not contain any tables.
            noLayers:
              summary: A `layers` request on a PSD with nothing drawable in it.
              value:
                error:
                  code: E_NO_LAYERS
                  message: This PSD file does not contain any layers with images that can be extracted.
            notTabular:
              summary: A `csv` request on a JSON document that is a single object, not a list.
              value:
                error:
                  code: E_NOT_TABULAR
                  message: This document is not a flat table of records (a top-level list of flat objects), so it cannot become CSV.
...
    RateLimited:
      description: |
        Too many requests from this address in the current window.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_RATE_LIMITED
              message: Too many requests. Try again in a moment.
    ConvertFailed:
      description: |
        The conversion produced nothing. Covers a damaged document, a document
        whose content does not match its extension, and a zero-byte upload.

        Note that LibreOffice's exit code is **not** used to decide this:
        `--convert-to` exits `0` even when it converts nothing, so the only
        trustworthy signal is whether a non-empty file of the expected type
        appeared on disk.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          # A 500 is either something LibreOffice refused or an unanticipated
          # fault. Both are documented, because a client that only handles the
          # first will be surprised by the second.
          examples:
            damaged:
              summary: The conversion produced no file
              value:
                error:
                  code: E_CONVERT_FAILED
                  message: This document could not be converted. It may be damaged or in a format the converter does not support.
            internal:
              summary: Unanticipated server fault
              value:
                error:
                  code: E_INTERNAL
                  message: Something went wrong on the server.
    Busy:
      description: |
        Every conversion slot is taken and the waiting queue is full. Returned
        immediately rather than queueing without bound, so the client gets a
        fast, honest answer instead of a request that hangs until it times out.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_BUSY
              message: The converter is busy. Try again in a moment.
    Timeout:
      description: |
        The conversion exceeded the server's 90 second deadline. The running
        LibreOffice process is sent `SIGTERM` and then `SIGKILL`.

        This is deliberately shorter than the client's own 120 second abort, so
        that the server answers with this response rather than being killed
        mid-conversion and leaving the phone to report a network failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: E_TIMEOUT
              message: This document took too long to convert.